Создание модулей, их связь, получение прогноза погоды (Вывод только названия города)

+++ data/src/main/java/ru/z3rg/data/remote/models/Day.kt
@@ -6,9 +6,9 @@ import com.google.gson.annotations.SerializedName
 data class Day(
 
     @SerializedName("avgtemp_c") var avgtempC: Double? = null,
-    @SerializedName("maxwind_kph") var maxwindKph: Int? = null,
-    @SerializedName("avghumidity") var avghumidity: Int? = null,
-    @SerializedName("daily_chance_of_rain") var dailyChanceOfRain: Int? = null,
+    @SerializedName("maxwind_kph") var maxwindKph: Double? = null,
+    @SerializedName("avghumidity") var avghumidity: Double? = null,
+    @SerializedName("daily_chance_of_rain") var dailyChanceOfRain: Double? = null,
     @SerializedName("condition") var condition: Condition? = Condition(),
 
     )
